Fix: HistologyWash mass field expects an int (#38)
* fix: histology wash mass defined as int

* linter fix

+++ aind-slims-service-server/src/aind_slims_service_server/models.py
@@ -125,7 +125,7 @@ class HistologyWashData(BaseModel):
     end_time: Optional[AwareDatetime] = None
     modified_by: Optional[str] = None
     reagents: List[HistologyReagentData] = []
-    mass: Optional[Decimal] = None
+    mass: Optional[int] = None
 
 
 class SlimsHistologyData(BaseModel):
